requirements.txt
flask~=3.0.3
requests~=2.32.3
python-dotenv~=0.21.1
flask_cors~=5.0.0
gunicorn~=23.0.0

pillow~=10.4.0
